feat(preset-mini): support decoration
