+++ bench/package.json
@@ -9,7 +9,7 @@
     "fs-extra": "^10.0.0",
     "local-pkg": "^0.4.0",
     "postcss": "^8.4.4",
-    "tailwindcss": "^3.0.0-alpha.2",
+    "tailwindcss": "^3.0.0",
     "unocss": "workspace:*",
     "vite": "^2.7.1",
     "vite-plugin-windicss": "^1.5.4",
