+++ package.json
@@ -18,7 +18,7 @@
   },
   "devDependencies": {
     "@antfu/eslint-config": "^0.12.2",
-    "@antfu/ni": "^0.11.0",
+    "@antfu/ni": "^0.11.1",
     "@iconify-json/carbon": "^1.0.12",
     "@iconify-json/logos": "^1.0.1",
     "@iconify-json/mdi": "^1.0.12",
@@ -45,7 +45,7 @@
     "@unocss/vite": "workspace:*",
     "@unocss/webpack": "workspace:*",
     "@vitejs/plugin-vue": "^1.10.2",
-    "@vueuse/core": "^7.1.2",
+    "@vueuse/core": "^7.2.2",
     "brotli-size": "^4.0.0",
     "bumpp": "^7.1.1",
     "codemirror": "^5.64.0",
@@ -61,16 +61,16 @@
     "splitpanes": "^3.0.6",
     "terser": "^5.10.0",
     "tsup": "5.10.3",
-    "typescript": "^4.5.2",
+    "typescript": "^4.5.3",
     "unbuild": "^0.6.3",
     "unocss": "workspace:*",
-    "unplugin-auto-import": "^0.5.1",
+    "unplugin-auto-import": "^0.5.3",
     "unplugin-vue-components": "^0.17.6",
     "vite": "^2.7.1",
     "vite-node": "^0.1.11",
     "vite-plugin-inspect": "^0.3.11",
-    "vite-plugin-pages": "^0.19.0-beta.6",
-    "vitest": "^0.0.39",
+    "vite-plugin-pages": "^0.19.0-beta.8",
+    "vitest": "^0.0.50",
     "vue": "^3.2.24",
     "vue-router": "^4.0.12"
   }
