+++ packages/preset-mini/src/rules/index.ts
@@ -19,3 +19,4 @@ export * from './transform'
 export * from './transition'
 export * from './typography'
 export * from './variables'
+export * from './decoration'
