+++ packages/preset-wind/src/rules/default.ts
@@ -48,11 +48,8 @@ import {
   tabSizes,
   textAligns,
   textColors,
-  textDecorationColors,
-  textDecorationLengths,
-  textDecorationOffsets,
-  textDecorationStyles,
   textDecorations,
+  textDecorationColors,
   textIndents,
   textOverflows,
   textShadows,
@@ -113,10 +110,7 @@ export const rules: Rule[] = [
   textIndents,
   textOverflows,
   textDecorations,
-  textDecorationStyles,
   textDecorationColors,
-  textDecorationLengths,
-  textDecorationOffsets,
   textStrokeWidths,
   textStrokeColors,
   textShadows,
