+++ packages/vscode/package.json
@@ -37,7 +37,7 @@
     "workspaceContains:package.json"
   ],
   "devDependencies": {
-    "@types/vscode": "^1.62.0",
+    "@types/vscode": "^1.63.0",
     "prettier": "^2.5.1",
     "tsup": "^5.10.3"
   },
