+++ test/fixtures/sveltekit/package.json
@@ -18,7 +18,7 @@
     "svelte-check": "^2.2.10",
     "svelte-preprocess": "^4.9.8",
     "tslib": "^2.3.1",
-    "typescript": "^4.5.2",
+    "typescript": "^4.5.3",
     "unocss": "link:../../../packages/unocss"
   },
   "type": "module"
