+++ tsconfig.json
@@ -16,6 +16,10 @@
       "@unocss/vite": ["./packages/vite/src/index.ts"],
       "@unocss/config": ["./packages/config/src/index.ts"],
       "@unocss/inspector": ["./packages/inspector/node/index.ts"],
+      "@unocss/preset-mini": ["./packages/preset-mini/src/index.ts"],
+      "@unocss/preset-mini/rules": ["./packages/preset-mini/src/rules.ts"],
+      "@unocss/preset-mini/variants": ["./packages/preset-mini/src/variants.ts"],
+      "@unocss/preset-mini/utils": ["./packages/preset-mini/src/utils.ts"],
       "@unocss/preset-uno": ["./packages/preset-uno/src/index.ts"],
       "@unocss/preset-icons": ["./packages/preset-icons/src/index.ts"],
       "@unocss/preset-attributify": ["./packages/preset-attributify/src/index.ts"],
